fixed js errors not so Chrome wouldn't run it
the dynamic loading with jinja2 caused some elements to not be loaded, thus causing the lines with js events tied to these elements to crash the rest of the code in Chrome, although if statements had been made to check if the content was loaded. Now instead I specifivally check if the elements have been rendered and that seems to work.